Move old middlewares to api/midddleware
