+++ api/api_tests.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/praelatus/backend/models"
-	"github.com/praelatus/backend/mw"
 	"github.com/praelatus/backend/store"
 )
