--- mw/mw.go
@@ -1,9 +1,0 @@
-package mw
-
-import "net/http"
-
-// Default is the default middleware stack for Praelatus
-var Default = []func(http.Handler) http.Handler{
-	Logger,
-	Auth,
-}
